Adjust to using a provider so we don't load powerlevels in our callcontainer
